Thin Particles Update
